+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed recursion from negative big number encoding
 - Expand test coverage of big number encoding
 - Source code nits in qcbor_number_encode
+- Improve QCBOREncode_OpenBytes() nesting error detection
 
 ### Fixed
 - Fix encoding of negative big nums that are one non-zero byte followed by zeros.